# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Crux API
  version: 0.1.0
  summary: Backend contract for Crux v1 (Phases 2-5).
  description: |
    Source of truth for Crux HTTP contracts. Pydantic models under
    `backend/contracts/` are generated from this document via
    `backend/scripts/gen_contracts.py`. Edit YAML, then regen.

    See `docs/superpowers/specs/2026-05-03-crux-v1-design.md`
    sections 3.3, 3.4, 4.1-4.4, 5 for upstream design constraints.
servers:
  - url: http://localhost:8000
    description: Local development backend
security:
  - BearerAuth: []
...
paths:
...
  /api/review/queue:
    get:
      tags: [review]
      summary: Concepts due for spaced-repetition review across all of the user's sessions.
      operationId: getReviewQueue
      parameters:
        - in: query
          name: limit
          required: false
          schema: { type: integer, default: 20, minimum: 1, maximum: 100 }
        - in: query
          name: offset
          required: false
          schema: { type: integer, default: 0, minimum: 0 }
      responses:
        "200":
          description: A page of due review concepts, most overdue first.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewQueuePage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/me:
    get:
      tags: [me]
      summary: Current user's account-level preferences and onboarding state.
      operationId: getMe
      responses:
        "200":
          description: Current user state (row auto-created on first call).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    patch:
      tags: [me]
      summary: Update display name, feedback preference, or onboarding state.
      operationId: patchMe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MePatchRequest"
      responses:
        "200":
          description: Updated user state.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Empty patch; at least one field must be provided.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
    delete:
      tags: [me]
      summary: Delete the caller's account, every row, stored upload, and the auth user.
      operationId: deleteMe
      responses:
        "204":
          description: Account deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >-
            A row for this account was written while the delete ran (for
            example a streaming reply or an upload still being processed).
            Nothing was deleted; retry once that work has settled.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: >-
            The auth admin service key is not configured (nothing was
            deleted), app data was removed but the auth user removal
            failed, or the auth JWKS endpoint is unavailable. The
            ErrorResponse detail names the failed step.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/me/export:
    get:
      tags: [me]
      summary: Download the caller's account, sessions, messages, check answers, uploads, and usage as one JSON file.
      operationId: exportMe
      description: >-
        Pure read. Uploads are listed by metadata only; the PDF bytes are not
        included. Internal state (embeddings, open-check pointers, per-call
        LLM logs) is left out. A caller with no users row gets an export with
        a null account and empty lists.
      responses:
        "200":
          description: The full export, sent as an attachment.
          headers:
            Content-Disposition:
              description: attachment; filename="crux-export-YYYY-MM-DD.json"
              schema: { type: string }
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataExport"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /api/documents/{document_id}:
    delete:
      tags: [documents]
      summary: Delete an uploaded reference document and its embeddings.
      operationId: deleteDocument
      parameters:
        - in: path
          name: document_id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Document deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Supabase-issued JWT. Server validates against Supabase JWKS and
        resolves the request's `user_id` from the `sub` claim. Replaces the
        Phase 6-and-earlier convention of passing `user_id` via body/query.
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    Unauthorized:
      description: Missing, invalid, or expired bearer token.
    ServiceUnavailable:
      description: Upstream dependency unavailable (auth JWKS outage or LLM provider).
...
  schemas:
...
    KnowledgeLevel:
      type: string
      enum: [beginner, intermediate, advanced]
...
    FeedbackPref:
      type: string
      enum: [hints, direct_answers]
    CheckIns:
      type: string
      enum: [often, sometimes, only_when_asked]
    ReplyLength:
      type: string
      enum: [brief, balanced, thorough]
...
    ConceptEvidence:
      type: string
      enum: [declared, tested]
      description: >
        Persisted provenance on a profile concept. "inferred" exists as tool-arg
        EvidenceType but is never persisted onto an entry.
    ConceptEntry:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        evidence_type:
          oneOf:
            - $ref: "#/components/schemas/ConceptEvidence"
            - type: "null"
          default: null
        last_event_at:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
          default: null
    TopicProfile:
      type: object
      additionalProperties: false
      description: |
        v2 (slice 8, 2026-07-12): concept lists carry ConceptEntry objects
        (previously plain strings) and subtopic_levels was added. Legacy
        string-element blobs are upgraded on read by the backend's tolerant
        parser; API consumers only ever see the v2 shape.
      properties:
        knowledge_level:
          oneOf:
            - $ref: "#/components/schemas/KnowledgeLevel"
            - type: "null"
          default: null
        subtopic_levels:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/KnowledgeLevel"
          default: {}
          description: >
            Per-subtopic knowledge levels keyed by canonicalized
            (strip+casefold) agent-named subtopic. knowledge_level is the
            session-wide default for subtopics without an entry.
        confirmed_gaps:
          type: array
          items: { $ref: "#/components/schemas/ConceptEntry" }
          default: []
        mastered_concepts:
          type: array
          items: { $ref: "#/components/schemas/ConceptEntry" }
          default: []
        focus_target_gap:
          type: [string, "null"]
          default: null
        last_session_summary:
          type: [string, "null"]
          default: null
...
    Citation:
      type: object
      additionalProperties: false
      required: [doc_id, text]
      properties:
        doc_id:   { type: string }
        text:     { type: string }
        page:     { type: [integer, "null"], default: null }
        doc_name: { type: [string, "null"], default: null }
...
    ReviewQueueItem:
      type: object
      additionalProperties: false
      required: [concept, source_session_id, source_topic, last_tested_at, streak, due_at]
      description: One concept due for spaced-repetition review.
      properties:
        concept: { type: string }
        source_session_id: { type: string }
        source_topic: { type: string }
        last_tested_at: { type: string, format: date-time }
        streak: { type: integer }
        due_at: { type: string, format: date-time }
    ReviewQueuePage:
      type: object
      additionalProperties: false
      required: [items, total, limit, offset]
      description: One page of the review queue, most overdue first.
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/ReviewQueueItem" }
        total:  { type: integer }
        limit:  { type: integer }
        offset: { type: integer }
...
    MeResponse:
      type: object
      additionalProperties: false
      required: [onboarding_complete, feedback_pref, check_ins, reply_length]
      properties:
        display_name:        { type: [string, "null"], default: null, maxLength: 120 }
        feedback_pref:
          allOf: [{ $ref: "#/components/schemas/FeedbackPref" }]
          default: hints
        check_ins:
          allOf: [{ $ref: "#/components/schemas/CheckIns" }]
          default: sometimes
        reply_length:
          allOf: [{ $ref: "#/components/schemas/ReplyLength" }]
          default: balanced
        onboarding_complete: { type: boolean, default: false }
    MePatchRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        display_name:        { type: string, maxLength: 120 }
        feedback_pref:       { $ref: "#/components/schemas/FeedbackPref" }
        check_ins:           { $ref: "#/components/schemas/CheckIns" }
        reply_length:        { $ref: "#/components/schemas/ReplyLength" }
        onboarding_complete: { type: boolean }

    # ---- Data export (#361) ----
    DataExport:
      type: object
      additionalProperties: false
      required: [format_version, exported_at, account, sessions, usage]
      description: Everything the server holds for one learner, oldest first.
      properties:
        format_version:
          type: integer
          description: Bumped when a field is removed or changes meaning.
        exported_at: { type: string, format: date-time }
        account:
          oneOf:
            - $ref: "#/components/schemas/ExportAccount"
            - type: "null"
        sessions:
          type: array
          items: { $ref: "#/components/schemas/ExportSession" }
        usage:
          type: array
          items: { $ref: "#/components/schemas/ExportUsageDay" }
    ExportAccount:
      type: object
      additionalProperties: false
      required: [user_id, created_at, feedback_pref, check_ins, reply_length, onboarding_complete]
      properties:
        user_id:             { type: string }
        created_at:          { type: string, format: date-time }
        display_name:        { type: [string, "null"], default: null }
        feedback_pref:       { $ref: "#/components/schemas/FeedbackPref" }
        check_ins:           { $ref: "#/components/schemas/CheckIns" }
        reply_length:        { $ref: "#/components/schemas/ReplyLength" }
        onboarding_complete: { type: boolean }
        accepted_terms_at:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
          default: null
        terms_version:       { type: [string, "null"], default: null }
    ExportSession:
      type: object
      additionalProperties: false
      required: [id, topic, created_at, pinned, topic_profile, messages, check_answers, documents]
      properties:
        id:         { type: string }
        topic:      { type: string }
        created_at: { type: string, format: date-time }
        ended_at:
          oneOf:
            - { type: string, format: date-time }
            - type: "null"
          default: null
        pinned:        { type: boolean }
        topic_profile: { $ref: "#/components/schemas/TopicProfile" }
        messages:
          type: array
          items: { $ref: "#/components/schemas/ExportMessage" }
        check_answers:
          type: array
          items: { $ref: "#/components/schemas/ExportCheckAnswer" }
        documents:
          type: array
          items: { $ref: "#/components/schemas/ExportDocument" }
    ExportMessage:
      type: object
      additionalProperties: false
      required: [id, role, content, created_at]
      properties:
        id:         { type: integer }
        role:       { type: string }
        content:    { type: string }
        created_at: { type: string, format: date-time }
        status:     { type: [string, "null"], default: null }
        citations:
          type: array
          items: { $ref: "#/components/schemas/Citation" }
          default: []
    ExportCheckAnswer:
      type: object
      additionalProperties: false
      required: [id, gap_tested, question, correct, created_at]
      properties:
        id:             { type: integer }
        gap_tested:     { type: string }
        question:       { type: string }
        correct:        { type: boolean }
        created_at:     { type: string, format: date-time }
        options:
          type: array
          items: { type: string }
          default: []
        selected_index: { type: [integer, "null"], default: null }
        correct_index:  { type: [integer, "null"], default: null }
        purpose:        { type: [string, "null"], default: null }
    ExportDocument:
      type: object
      additionalProperties: false
      required: [id, filename, created_at]
      description: Upload metadata only; the file itself is not in the export.
      properties:
        id:         { type: integer }
        filename:   { type: string }
        status:     { type: [string, "null"], default: null }
        error:      { type: [string, "null"], default: null }
        page_count: { type: [integer, "null"], default: null }
        created_at: { type: string, format: date-time }
    ExportUsageDay:
      type: object
      additionalProperties: false
      required: [date_utc, messages, cost_usd]
      properties:
        date_utc: { type: string, format: date }
        messages: { type: integer, minimum: 0 }
        cost_usd: { type: number, minimum: 0 }
...
    CodedErrorDetail:
      type: object
      additionalProperties: true
      required: [code]
      properties:
        code: { type: string }
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [detail]
      properties:
        detail:
          oneOf:
            - type: string
            - $ref: "#/components/schemas/CodedErrorDetail"

# ---------------------------------------------------------------------------
# SSE event contract reference (vendor extension — not processed by codegen)
# Describes the Server-Sent Events emitted by POST /chat/stream.
# Field names must match what agent/tutor.py actually yields.
# ---------------------------------------------------------------------------
...